refactor components - pages
