+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './Footer'
 export * from './NavBar'
-export * from './Schedule'
