+++ src/components/news/index.ts
@@ -1,0 +1,1 @@
+export * from './CardNews'
