+++ src/components/schedule/index.ts
