+++ src/pages/index.ts
@@ -2,4 +2,4 @@ export * from './about'
 export * from './home'
 export * from './schedule'
 export * from './news'
-export * from './routes'
+export * from '../routes'
