--- src/pages/news/index.ts
@@ -1,1 +1,0 @@
-export * from './News'
